`evaluation` - `load_data_from_excel` moved to `excel_sheet_utils`
